+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.18.1] - 2025-09-01
+
 ### Added
 
 - Added a component `ExternallyMocked` to exclude actions from being updated. These actions are updated manually by the user.
@@ -414,7 +416,8 @@ This update features a big rewrite into a component-based API. The core concepts
 
 Initial release.
 
-[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.18.0...HEAD
+[unreleased]: https://github.com/simgine/bevy_replicon/compare/v0.18.1...HEAD
+[0.18.1]: https://github.com/simgine/bevy_replicon/compare/v0.18.0...v0.18.1
 [0.18.0]: https://github.com/simgine/bevy_replicon/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/simgine/bevy_replicon/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/simgine/bevy_replicon/compare/v0.15.3...v0.16.0